requirements.txt
slackclient
requests
pyaml
datetime
